+++ R/tm_t_abnormality_by_worst_grade.R
@@ -58,8 +58,13 @@ template_abnormality_by_worst_grade <- function(parentname, #nolint
                               #Changed the following prepo step methodology as not
                               #all cases have grade = 4 (realized with nsdl real data)
                               GRADE_ANL = factor(
-                                abs(as.numeric(as.character(atoxgr_var)))
-                                )) %>%
+                                abs(
+                                  as.numeric(
+                                    as.character(atoxgr_var)
+                                    )
+                                  )
+                                )
+                              ) %>%
                             filter(WGRLOFL == TRUE | WGRHIFL == TRUE) %>%
                             droplevels(),
                           env  = list(
@@ -74,7 +79,9 @@ template_abnormality_by_worst_grade <- function(parentname, #nolint
 
   data_list <- add_expr(
     data_list,
-    quote(expr = var_labels(anl) <- c(anl_labels, "GRADE_DIR", "GRADE_ANL"))
+    quote(
+      expr = var_labels(anl) <- c(anl_labels, "GRADE_DIR", "GRADE_ANL")
+      )
     )
 
   data_list <- add_expr(