+++ NEWS.md
@@ -1,5 +1,7 @@
 # teal.modules.clinical 0.8.10.9000
 
+* Updated R version requirement to >= 3.6.
+
 # teal.modules.clinical 0.8.10
 ### New features
 * New module `tm_t_smq` was added for the analysis of adverse events by Standardized `MedDRA` Query.