+++ NEWS.md
@@ -24,6 +24,7 @@
 * Used `format_count_fraction` to fix formatting inconsistency in `tm_t_events_summary`.
 * Updated `count_occurrences` `vars` argument in `tm_t_shift_by_grade`.
 * Updated `tm_t_pp_laboratory` to display 4 decimals by default.
+* Updated `tm_t_events_by_grade` to use the `trim_levels_in_group` split function instead of the `trim_rows` function.
 
 ### Miscellaneous
 * Updated `LICENCE` and `README` with new package references.