+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.modules.clinical 0.8.10.9002
+# teal.modules.clinical 0.8.10.9003
 
 * Updated R version requirement to >= 3.6.
 * Updated `tm_t_abnormality` due to changes in `count_abnormal` that `abnormal` argument is taking list as input now.